requirements/mydre_uploader.txt
-r base.txt
Pillow>=10.0.0
requests>=2.31.0
azure-storage-blob>=12.0.0 